정거장 튜토리얼 준비

+++ Train/Assets/_Script/Train/Medic_Train.cs
@@ -9,7 +9,7 @@ public class Medic_Train : MonoBehaviour
     public GameObject[] On_Off_Object;
 
     int Heal_Amount;
-    int Heal_timeBet;
+    float Heal_timeBet;
     bool isPlayerHealing;
     public bool isMercenaryHealing;
     bool isHeal;
